slack/requirements.txt
pandas>=1.3.0
python-dotenv>=0.19.0
slack-sdk>=3.20.0
tqdm>=4.60.0